codestyle changes + calendar replaced with Date
